rename: 커스텀 클래스 및 프로토콜을 독립된 파일로 분리
