+++ JuiceMaker/JuiceMaker/Controller/JuiceOrderViewControllerDelegate.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+protocol JuiceOrderViewControllerDelegate: AnyObject {
+    func JuiceOrderViewControllerHasChanges()
+}
