+++ JuiceMaker/JuiceMaker/View/Class/FruitLabel.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+class FruitLabel: UILabel {
+    var fruitType: FruitType?
+}
