+++ JuiceMaker/JuiceMaker/View/Class/FruitStepper.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+class FruitStepper: UIStepper {
+    var fruitType: FruitType?
+}
