+++ JuiceMaker/JuiceMaker/View/Class/JuiceButton.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+class JuiceButton: UIButton {
+    var juiceType: JuiceMaker.JuiceType?
+}
